New enhancement on supporting clustering on large dataset (test on about 70k samples); verbose for running information; passing density peaks for clustering
fix errors in c++ and test

update the cluster assignment

add the argument peaks to findClusters.densityCluster

update documentation findClusters.densityCluster

prepare for pull request

prepare for pull request

prepare for pull request

update the Rd files

update scripts based on Thomas's suggestions

+++ src/distanceToPeak.cpp
@@ -3,13 +3,13 @@ using namespace Rcpp;
 
 // [[Rcpp::export]]
 NumericVector distanceToPeakCpp(NumericVector distance, NumericVector rho) {
-   int size = rho.size();
+   size_t size = rho.size();
    NumericVector peaks(size);
    NumericVector maximum(size);
    
-   int i = 0;
-   for (int col = 0; col < size; col++) {
-      for (int row = col + 1; row < size; row++) {
+   size_t i = 0;
+   for (size_t col = 0; col < size; col++) {
+      for (size_t row = col + 1; row < size; row++) {
          double newValue = distance[i];
          double rhoRow = rho[row];
          double rhoCol = rho[col];
@@ -35,7 +35,7 @@ NumericVector distanceToPeakCpp(NumericVector distance, NumericVector rho) {
       }
    }
    
-   for (int j = 0; j < size; j++) {
+   for (size_t j = 0; j < size; j++) {
       if (peaks[j] == 0) {
          peaks[j] = maximum[j];
       } else {
